requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4

# ETL & Analytics
prefect==2.14.21
polars==0.20.2
pandas==2.1.4
pyarrow==15.0.0
httpx==0.26.0
requests==2.31.0
python-levenshtein==0.23.0
rapidfuzz==3.6.1
scikit-learn==1.4.0
statsmodels==0.14.1
duckdb==0.10.0

# PDF & Billing
reportlab==4.0.9
stripe==7.11.0

# Frontend (Streamlit)
streamlit==1.31.0
plotly==5.18.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Dev
ruff==0.1.11
black==24.1.1
mypy==1.8.0
pre-commit==3.6.0

